Added the details about population evidence by using various access modifiers.
